LayoutTests/canvas/philip/tests/size.attributes.get-expected.txt

< [index] >
size.attributes.get
Getting width/height IDL/content attributes
References: size.width size.height
Actual output:
Expected output:

Passed
